+++ src/main/scala/com/raquo/laminar/nodes/ChildNode.scala
@@ -36,7 +36,7 @@ trait ChildNode[+Ref <: dom.Node]
     *
     * @param maybeNextParent  `None` means this node is about to be detached form its parent
     */
-  @inline def willSetParent(maybeNextParent: Option[ParentNode.Base]): Unit = {}
+  @inline def willSetParent(maybeNextParent: Option[ParentNode.Base]): Unit = ()
 
   override def apply(parentNode: ParentNode.Base): Unit = {
     // @TODO[Performance] Consider making ChildNode -> Modifier conversion implicit instead (but watch compile times)