+++ src/main/scala/com/raquo/laminar/nodes/ReactiveElement.scala
@@ -1,15 +1,15 @@
 package com.raquo.laminar.nodes
 
-import com.raquo.airstream.core.{Observable, Observer, Subscription}
-import com.raquo.airstream.eventbus.{EventBus, EventBusSource, WriteBus}
+import com.raquo.airstream.core.{Observable, Observer}
+import com.raquo.airstream.eventbus.{EventBus, WriteBus}
 import com.raquo.airstream.eventstream.EventStream
-import com.raquo.airstream.ownership.{Owned, Owner}
+import com.raquo.airstream.ownership.{DynamicOwner, DynamicSubscription, Owner, Subscription}
 import com.raquo.airstream.signal.Signal
 import com.raquo.domtypes
 import com.raquo.domtypes.generic.keys.EventProp
 import com.raquo.laminar.DomApi
 import com.raquo.laminar.emitter.EventPropEmitter
-import com.raquo.laminar.lifecycle.{MountEvent, NodeDidMount, NodeWasDiscarded, NodeWillUnmount, ParentChangeEvent}
+import com.raquo.laminar.lifecycle.{MountEvent, NodeDidMount, NodeWillUnmount, ParentChangeEvent}
 import com.raquo.laminar.nodes.ChildNode.isParentMounted
 import com.raquo.laminar.nodes.ReactiveElement.noMountEvents
 import com.raquo.laminar.receivers.{ChildReceiver, ChildrenCommandReceiver, ChildrenReceiver, MaybeChildReceiver, TextChildReceiver}
@@ -21,35 +21,35 @@ import scala.collection.mutable
 trait ReactiveElement[+Ref <: dom.Element]
   extends ChildNode[Ref]
   with ParentNode[Ref]
-  with domtypes.generic.nodes.Element
-  with Owner {
+  with domtypes.generic.nodes.Element {
 
   // @TODO[Naming] We reuse EventPropSetter to represent an active event listener. Makes for a bit confusing naming.
   private[ReactiveElement] var _maybeEventListeners: Option[mutable.Buffer[EventPropSetter[_]]] = None
 
   @inline def maybeEventListeners: Option[List[EventPropSetter[_]]] = _maybeEventListeners.map(_.toList)
 
+  private[this] val dynamicOwner: DynamicOwner = new DynamicOwner
+
   /** Event bus that emits parent change events.
     * For efficiency, it is only populated when someone accesses [[parentChangeEvents]]
     */
-  private[laminar] var maybeParentChangeBus: Option[EventBus[ParentChangeEvent]] = None
+  private[this] var maybeParentChangeBus: Option[EventBus[ParentChangeEvent]] = None
 
   /** Event bus that emits this node's mount events.
     * For efficiency, it is only populated when someone accesses [[thisNodeMountEvents]]
     */
-  private[laminar] var maybeThisNodeMountEventBus: Option[EventBus[MountEvent]] = None
+  private[this] var maybeThisNodeMountEventBus: Option[EventBus[MountEvent]] = None
 
   /** Stream of parent change events for this node.
-    * For efficiency, it is lazy loaded, only being initialized when accessed,
-    * either directly or (more commonly) as a dependency of [[mountEvents]]
+    * For efficiency, it is lazy loaded, only being initialized when accessed
     */
-  lazy val parentChangeEvents: EventStream[ParentChangeEvent] = {
+  private[this] lazy val parentChangeEvents: EventStream[ParentChangeEvent] = {
     val parentChangeBus = new EventBus[ParentChangeEvent]
     maybeParentChangeBus = Some(parentChangeBus)
     parentChangeBus.events
   }
 
-  lazy val maybeParentSignal: Signal[Option[ParentNode.Base]] = parentChangeEvents
+  private[this] lazy val maybeParentSignal: Signal[Option[ParentNode.Base]] = parentChangeEvents
     .filter(_.alreadyChanged) // @TODO[Integrity] is this important?
     .map(_.maybeNextParent)
     .toSignal(maybeParent)
@@ -57,7 +57,7 @@ trait ReactiveElement[+Ref <: dom.Element]
   /** Emits mount events that were caused by this element's parent changing its parent,
     * or any such changes up the chain. Does not include mount events triggered by this
     * element changing its parent - see [[thisNodeMountEvents]] for that. */
-  private lazy val ancestorMountEvents: EventStream[MountEvent] = {
+  private[this] lazy val ancestorMountEvents: EventStream[MountEvent] = {
     maybeParentSignal.map {
       case Some(nextParent: ReactiveElement[_]) =>
         nextParent.mountEvents
@@ -69,7 +69,7 @@ trait ReactiveElement[+Ref <: dom.Element]
   /** Emits mount events caused by this node changing its parent. Does not include mount
     * events triggered by changes higher in the hierarchy (grandparent and up) – see
     * [[ancestorMountEvents]] for that. */
-  private lazy val thisNodeMountEvents: EventStream[MountEvent] = {
+  private[this] lazy val thisNodeMountEvents: EventStream[MountEvent] = {
     val thisNodeMountEventBus = new EventBus[MountEvent]
     maybeThisNodeMountEventBus = Some(thisNodeMountEventBus)
     thisNodeMountEventBus.events
@@ -78,6 +78,25 @@ trait ReactiveElement[+Ref <: dom.Element]
   /** Emits mount events for this node, including mount events fired by all of its ancestors */
   lazy val mountEvents: EventStream[MountEvent] = {
     EventStream.merge(ancestorMountEvents, thisNodeMountEvents)
+  }
+
+  // @nc normally this would not work memory management wise, as we create a sub on element init, and never even bother to kill it
+  //  - however, this sub only looks at mount events
+  //  - thisNodeMountEvents are contained within this node and don't require any leaky resources
+  //  - ancestorMountEvents look at parent nodes' mountEvents
+  //    - while this node is mounted, all ancestors are also mounted and that is ok
+  //    - if this node becomes unmounted, the ancestor chain includes only other unmounted nodes
+  //    - so this will lock together that ancestor chain, but that's like meh
+  //    - even if we re-mounted this node to different parents / ancestors, this chain would be broken as `mountEvents` would now point elsewhere
+  //    - as long as we understand and document this limitation, I think it's ok
+  // @nc Can we initialize this conditionally, only for elements that need it?
+  {
+    val pilotSubscriptionOwner: Owner = new Owner {}
+
+    mountEvents.foreach{
+      case NodeDidMount => dynamicOwner.activate()
+      case NodeWillUnmount => dynamicOwner.deactivate()
+    }(pilotSubscriptionOwner)
   }
 
   /** Create and get a stream of events on this node */
@@ -108,45 +127,42 @@ trait ReactiveElement[+Ref <: dom.Element]
 
   final def <--[V](childrenCommandReceiver: ChildrenCommandReceiver.type): ChildrenCommandReceiver = new ChildrenCommandReceiver(this)
 
-  // @TODO[Naming] Not a fan of `subscribeO` name, but it needs to be different from `subscribe` for type inference to work
+  // @TODO[Naming] Not a fan of `subscribeS` / `subscribeO` names, but it needs to be different from `subscribe` for type inference to work
+  // @TODO[API] Consider having subscribe() return a Subscribe object that has several apply methods on it to reign in this madness
+  //  - Also, do we really need currying here?
 
-  def subscribeO[V](observable: Observable[V])(observer: Observer[V]): Subscription = {
-    observable.addObserver(observer)(owner = this)
+  def subscribeS(getSubscription: Owner => Subscription): DynamicSubscription = {
+    new DynamicSubscription(dynamicOwner, getSubscription)
+  }
+
+  def subscribeO[V](observable: Observable[V])(observer: Observer[V]): DynamicSubscription = {
+    subscribeS(owner => observable.addObserver(observer)(owner))
   }
 
   // @TODO[Naming] Rename to subscribeThisO?
-  def subscribeO[V](getObservable: this.type => Observable[V])(observer: Observer[V]): Subscription = {
+  def subscribeO[V](getObservable: this.type => Observable[V])(observer: Observer[V]): DynamicSubscription = {
     subscribeO(getObservable(this))(observer)
   }
 
-  def subscribe[V](observable: Observable[V])(onNext: V => Unit): Subscription = {
+  def subscribe[V](observable: Observable[V])(onNext: V => Unit): DynamicSubscription = {
     subscribeO(observable)(Observer(onNext))
   }
 
   // @TODO[Naming] Rename to subscribeThis?
-  def subscribe[V](getObservable: this.type => Observable[V])(onNext: V => Unit): Subscription = {
+  def subscribe[V](getObservable: this.type => Observable[V])(onNext: V => Unit): DynamicSubscription = {
     subscribeO(getObservable(this))(Observer(onNext))
   }
 
   def subscribeBus[V](
     sourceStream: EventStream[V],
     targetBus: WriteBus[V]
-  ): EventBusSource[V] = {
-    targetBus.addSource(sourceStream)(owner = this)
-  }
-
-  /** Check whether the node is currently mounted.
-    *
-    * You can use this method to simplify your code and possibly improve performance
-    * where you'd otherwise need to subscribe to and transform [[mountEvents]]
-    */
-  def isMounted: Boolean = {
-    isParentMounted(maybeParent)
+  ): DynamicSubscription = {
+    subscribeS(owner => targetBus.addSource(sourceStream)(owner))
   }
 
   /** This hook is exposed by Scala DOM Builder for this exact purpose */
   override def willSetParent(maybeNextParent: Option[ParentNode.Base]): Unit = {
-    // @TODO[Integrity] In dombuilder, willSetParent is called at incorrect time in ParentNode.appendChild
+    // super.willSetParent(maybeNextParent) // default implementation is a noop, so no need to call it.
     // dom.console.log(">>>> willSetParent", this.ref.tagName + "(" + this.ref.textContent + ")", maybeParent == maybeNextParent, maybeParent.toString, maybeNextParent.toString)
     if (maybeNextParent != maybeParent) {
       maybeParentChangeBus.foreach(bus => {
@@ -171,37 +187,21 @@ trait ReactiveElement[+Ref <: dom.Element]
     super.setParent(maybeNextParent)
     // dom.console.log(">>>> setParent", this.ref.tagName + "(" + this.ref.textContent + ")", maybePrevParent == maybeNextParent, maybePrevParent.toString, maybeNextParent.toString)
     if (maybeNextParent != maybePrevParent) {
-      maybeParentChangeBus.foreach( bus => {
-        bus.writer.onNext(ParentChangeEvent(
+      maybeParentChangeBus.foreach(bus => {
+        val ev = ParentChangeEvent(
           alreadyChanged = true,
           maybePrevParent = maybePrevParent,
           maybeNextParent = maybeNextParent
-        ))
+        )
+        bus.writer.onNext(ev)
       })
       val prevParentIsMounted = isParentMounted(maybePrevParent)
       val nextParentIsMounted = isParentMounted(maybeNextParent)
       if (!prevParentIsMounted && nextParentIsMounted) {
         maybeThisNodeMountEventBus.foreach(_.writer.onNext(NodeDidMount))
       }
-      if (prevParentIsMounted && !nextParentIsMounted) {
-        maybeThisNodeMountEventBus.foreach { bus =>
-          bus.writer.onNext(NodeWasDiscarded) // @TODO this should be optional
-        }
-      }
     }
   }
-
-  override protected[this] def onOwned(owned: Owned): Unit = {
-    val isFirstSubscription = possessions.length == 1
-    if (isFirstSubscription) {
-      // Create a subscription that will kill all subscriptions that this element owns when it is discarded.
-      // Note: Once this subscription is created, it will persist until this element is discarded.
-      //       This might be suboptimal in a narrow range of conditions, I don't think anyone will run into those.
-      // @TODO[Performance] ^^^^ The above can be addressed by e.g. overriding onKilledExternally (but permissions)
-      mountEvents.filter(_ == NodeWasDiscarded).foreach(_ => killPossessions() )(owner = this)
-    }
-  }
-
 }
 
 object ReactiveElement {